requirements.txt
appdirs==1.4.4
backcall==0.2.0
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
freetype-py==2.2.0
imageio==2.9.0
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
kiwisolver==1.2.0
llvmlite==0.34.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mesh-to-sdf==0.0.12
meshio==4.3.0
networkx==2.5
numba==0.51.2
numpy==1.19.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycuda==2020.1
pyglet==1.5.8
Pygments==2.7.1
pynrrd==0.4.2
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrender==0.1.43
python-dateutil==2.8.1
pytools==2020.4
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.9.3
tqdm==4.50.2
traitlets==4.3.3
trimesh==3.8.11
wcwidth==0.2.5
zipp==3.3.0
